backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-SocketIO
python-dotenv
huey[sqlite]
Flask-Cors
gunicorn
eventlet
rarfile
py7zr
Pillow